Add listener registration to PlotAPI. Add documentation and refactor Result enum
